Переместил структуру `AesCipherParams` в cpp-файл
